Rewritten with a simplified python-six module beatbox_six
