+++ xmltramp.py
@@ -1,8 +1,7 @@
 """xmltramp: Make XML documents easily accessible."""
 
 import re
-from six import PY2, StringIO, text_type
-from six.moves import xrange
+from beatbox_six import python_2_unicode_compatible, StringIO, text_type, xrange
 from xml.sax.handler import EntityResolver, DTDHandler, ContentHandler, ErrorHandler
 from xml.sax import make_parser
 from xml.sax.handler import feature_namespaces
@@ -33,24 +32,6 @@ def quote(x, elt=True):
     if not elt:
         x = x.replace('"', '&quot;')
     return x
-
-
-def python_2_unicode_compatible(klass):
-    """
-    A decorator that defines __unicode__ and __str__ methods under Python 2.
-    Under Python 3 it does nothing.
-
-    To support Python 2 and 3 with a single code base, define a __str__ method
-    returning text and apply this decorator to the class.
-    """
-    if PY2:
-        if '__str__' not in klass.__dict__:
-            raise ValueError("@python_2_unicode_compatible cannot be applied "
-                             "to %s because it doesn't define __str__()." %
-                             klass.__name__)
-        klass.__unicode__ = klass.__str__
-        klass.__str__ = lambda self: self.__unicode__().encode('utf-8')
-    return klass
 
 
 @python_2_unicode_compatible
